Move the locks module to perun core and fix some codestyle issues.
