+++ codecov.yml
@@ -1,4 +1,4 @@
 ignore: 
   - "perun/collect/trace"
-  - "tests/test_collect"
+  - "tests/test_tracer"
